split StringExtensions into 2 parts
