+++ packages/http-client-csharp/generator/Microsoft.TypeSpec.Generator.ClientModel/src/Providers/ClientOptionsProvider.cs
@@ -7,7 +7,7 @@ using System.IO;
 using System.Linq;
 using Microsoft.TypeSpec.Generator.Expressions;
 using Microsoft.TypeSpec.Generator.Input;
-using Microsoft.TypeSpec.Generator.Input.Utilities;
+using Microsoft.TypeSpec.Generator.Input.Extensions;
 using Microsoft.TypeSpec.Generator.Primitives;
 using Microsoft.TypeSpec.Generator.Providers;
 using Microsoft.TypeSpec.Generator.Utilities;
@@ -140,7 +140,7 @@ namespace Microsoft.TypeSpec.Generator.ClientModel.Providers
                             description,
                             MethodSignatureModifiers.Public,
                             type,
-                            p.Name.ToCleanIdentifierName(),
+                            p.Name.ToIdentifierName(),
                             new AutoPropertyBody(true),
                             this));
                     }
