+++ packages/http-client-csharp/generator/Microsoft.TypeSpec.Generator.ClientModel/test/Providers/ScmMethodProviderCollectionTests.cs
@@ -8,7 +8,7 @@ using System.Linq;
 using System.Threading;
 using Microsoft.TypeSpec.Generator.ClientModel.Providers;
 using Microsoft.TypeSpec.Generator.Input;
-using Microsoft.TypeSpec.Generator.Input.Utilities;
+using Microsoft.TypeSpec.Generator.Input.Extensions;
 using Microsoft.TypeSpec.Generator.Primitives;
 using Microsoft.TypeSpec.Generator.Snippets;
 using Microsoft.TypeSpec.Generator.Tests.Common;
@@ -45,7 +45,7 @@ namespace Microsoft.TypeSpec.Generator.ClientModel.Tests.Providers
             var signature = method.Signature;
             Assert.IsNotNull(signature);
             var operation = serviceMethod.Operation;
-            Assert.AreEqual(operation.Name.ToCleanIdentifierName(), signature.Name);
+            Assert.AreEqual(operation.Name.ToIdentifierName(), signature.Name);
 
             var parameters = signature.Parameters;
             Assert.IsNotNull(parameters);
@@ -53,7 +53,7 @@ namespace Microsoft.TypeSpec.Generator.ClientModel.Tests.Providers
 
             var convenienceMethod = methodCollection.FirstOrDefault(m
                 => !m.Signature.Parameters.Any(p => p.Name == "content")
-                    && m.Signature.Name == $"{operation.Name.ToCleanIdentifierName()}");
+                    && m.Signature.Name == $"{operation.Name.ToIdentifierName()}");
             Assert.IsNotNull(convenienceMethod);
             Assert.AreEqual(serviceMethod, convenienceMethod!.ServiceMethod);
 
@@ -88,7 +88,7 @@ namespace Microsoft.TypeSpec.Generator.ClientModel.Tests.Providers
             var operation = serviceMethod.Operation;
             var asyncConvenienceMethod = methodCollection.FirstOrDefault(m
                 => !m.Signature.Parameters.Any(p => p.Name == "content")
-                    && m.Signature.Name == $"{operation.Name.ToCleanIdentifierName()}Async");
+                    && m.Signature.Name == $"{operation.Name.ToIdentifierName()}Async");
             Assert.IsNotNull(asyncConvenienceMethod);
 
             var asyncConvenienceMethodParameters = asyncConvenienceMethod!.Signature.Parameters;
@@ -101,7 +101,7 @@ namespace Microsoft.TypeSpec.Generator.ClientModel.Tests.Providers
 
             var syncConvenienceMethod = methodCollection.FirstOrDefault(m
                 => !m.Signature.Parameters.Any(p => p.Name == "content")
-                   && m.Signature.Name == operation.Name.ToCleanIdentifierName());
+                   && m.Signature.Name == operation.Name.ToIdentifierName());
             Assert.IsNotNull(syncConvenienceMethod);
 
             var syncConvenienceMethodParameters = syncConvenienceMethod!.Signature.Parameters;
