+++ packages/http-client-csharp/generator/Microsoft.TypeSpec.Generator.Input/src/InputTypes/Serialization/TypeSpecSerialization.cs
@@ -8,7 +8,7 @@ using System.Text;
 using System.Text.Json;
 using System.Text.Json.Serialization;
 using AutoRest.CSharp.Common.Input;
-using Microsoft.TypeSpec.Generator.Input.Utilities;
+using Microsoft.TypeSpec.Generator.Input.Extensions;
 
 namespace Microsoft.TypeSpec.Generator.Input
 {
@@ -78,7 +78,7 @@ namespace Microsoft.TypeSpec.Generator.Input
         {
             foreach (var client in clients)
             {
-                var cleanName = client.Name.ToCleanIdentifierName();
+                var cleanName = client.Name.ToIdentifierName();
                 client.Name = BuildClientName(cleanName, parentNames);
 
                 var lastSegment = GetLastSegment(client.Namespace);
