+++ packages/http-client-csharp/generator/Microsoft.TypeSpec.Generator/src/Utilities/FormattableStringHelpers.cs
@@ -8,8 +8,9 @@ using System.Diagnostics.CodeAnalysis;
 using System.Linq;
 using System.Runtime.CompilerServices;
 using System.Text;
-using Microsoft.TypeSpec.Generator.Input.Utilities;
+using Microsoft.TypeSpec.Generator.Input.Extensions;
 using Microsoft.TypeSpec.Generator.Providers;
+using Microsoft.TypeSpec.Generator.Utilities;
 
 namespace Microsoft.TypeSpec.Generator
 {
@@ -131,7 +132,7 @@ namespace Microsoft.TypeSpec.Generator
             // For instance, when the format string is all \n, it will produce n+1 segments (because we did not omit empty entries).
             Span<Range> splitIndices = stackalloc Range[input.Format.Length + 1];
             ReadOnlySpan<char> formatSpan = input.Format.AsSpan();
-            foreach ((ReadOnlySpan<char> span, bool isLiteral, int index) in StringHelpers.GetFormattableStringFormatParts(formatSpan))
+            foreach ((ReadOnlySpan<char> span, bool isLiteral, int index) in InternalStringExtensions.GetFormattableStringFormatParts(formatSpan))
             {
                 // if isLiteral - put in formatBuilder
                 if (isLiteral)
