+++ packages/http-client-csharp/generator/Microsoft.TypeSpec.Generator/test/Providers/PropertyProviderTests.cs
@@ -4,7 +4,7 @@
 using System.Collections.Generic;
 using System.Linq;
 using Microsoft.TypeSpec.Generator.Input;
-using Microsoft.TypeSpec.Generator.Input.Utilities;
+using Microsoft.TypeSpec.Generator.Input.Extensions;
 using Microsoft.TypeSpec.Generator.Primitives;
 using Microsoft.TypeSpec.Generator.Providers;
 using Microsoft.TypeSpec.Generator.Tests.Common;
@@ -81,7 +81,7 @@ namespace Microsoft.TypeSpec.Generator.Tests.Providers
             InputFactory.Model("TestModel", properties: [collectionProperty]);
             var property = new PropertyProvider(collectionProperty, new TestTypeProvider());
 
-            Assert.AreEqual(collectionProperty.Name.ToCleanIdentifierName(), property.Name);
+            Assert.AreEqual(collectionProperty.Name.ToIdentifierName(), property.Name);
             Assert.AreEqual(expectedType, property.Type);
 
             // validate the parameter conversion
@@ -122,7 +122,7 @@ namespace Microsoft.TypeSpec.Generator.Tests.Providers
             InputFactory.Model("TestModel", properties: [inputModelProperty]);
 
             var property = new PropertyProvider(inputModelProperty, testTypeProvider);
-            Assert.AreEqual(inputPropertyName.ToCleanIdentifierName() + "Property", property.Name);
+            Assert.AreEqual(inputPropertyName.ToIdentifierName() + "Property", property.Name);
         }
 
         [Test]
